soonho/logout - v1.0_로그아웃 버그 수정
